+++ backend/src/main/java/com/example/backend/master/platform/tenantDatabases/TenantDatabasesController.java
@@ -26,9 +26,9 @@ public class TenantDatabasesController {
 
     @CrossOrigin(origins = "*", allowedHeaders = "*")
     @GetMapping("/{id}")
-    public ResponseEntity<?> getById(@PathVariable(value = "id") Integer id){
+    public ResponseEntity<?> getById(@PathVariable(value = "id") Long id){
 
-        Optional<TenantDatabases> tenantDatabases = repository.findById(Long.valueOf(id));
+        Optional<TenantDatabases> tenantDatabases = repository.findById(id);
         if(tenantDatabases.isEmpty()){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não encontrado");
         }
@@ -47,9 +47,9 @@ public class TenantDatabasesController {
 
     @CrossOrigin(origins = "*", allowedHeaders = "*")
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateTenantDatabases(@PathVariable(value = "id") Integer id, @RequestBody TenantDatabasesRequestDTO upData){
+    public ResponseEntity<?> updateTenantDatabases(@PathVariable(value = "id") Long id, @RequestBody TenantDatabasesRequestDTO upData){
 
-        Optional<TenantDatabases> tenantDatabases = repository.findById(Long.valueOf(id));
+        Optional<TenantDatabases> tenantDatabases = repository.findById(id);
         if(tenantDatabases.isEmpty()){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não encontrado");
         }
@@ -61,9 +61,9 @@ public class TenantDatabasesController {
 
     @CrossOrigin(origins = "*", allowedHeaders = "*")
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteTenantDatabases(@PathVariable(value = "id") Integer id){
+    public ResponseEntity<?> deleteTenantDatabases(@PathVariable(value = "id") Long id){
 
-        Optional<TenantDatabases> tenantDatabases = repository.findById(Long.valueOf(id));
+        Optional<TenantDatabases> tenantDatabases = repository.findById(id);
         if(tenantDatabases.isEmpty()){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não encontrado");
         }
